+++ CharacterDataEditor/Enums/SpriteType.cs
@@ -1,0 +1,9 @@
+namespace CharacterDataEditor.Enums
+{
+    public enum SpriteType
+    {
+        HurtSprite,
+        GrabSprite,
+        KnockdownSprite
+    }
+}