+++ CharacterDataEditor/Screens/EditCharacterScreen.cs
@@ -36,6 +36,7 @@ namespace CharacterDataEditor.Screens
         private PaletteModel paletteInEditor;
         private List<string> moveTypesList = new List<string>();
         private List<string> attackTypesList = new List<string>();
+        private List<string> spriteTypesList = new List<string>();
         private List<SpriteDataModel> allSprites;
 
         private string spriteToDraw;
@@ -128,6 +129,16 @@ namespace CharacterDataEditor.Screens
                 var itemAsString = item.ToString().AddSpacesToCamelCase();
 
                 attackTypesList.Add(itemAsString);
+            }
+
+            var spriteTypes = Enum.GetValues(typeof(SpriteType));
+            spriteTypesList = new List<string>();
+
+            foreach (SpriteType item in spriteTypes)
+            {
+                var itemAsString = item.ToString().AddSpacesToCamelCase();
+
+                spriteTypesList.Add(itemAsString);
             }
 
             //init spritedrawposition
@@ -405,22 +416,13 @@ namespace CharacterDataEditor.Screens
             else
             {
                 int selectedMoveType = (int)moveInEditor.MoveType;
-
-                ImGui.Text("Move Type");
-                ImGui.SameLine();
-
-                ImGui.Combo("##MoveType", ref selectedMoveType, moveTypesList.ToArray(), moveTypesList.Count);
-
+                ImguiDrawingHelper.DrawComboInput("moveType", moveTypesList.ToArray(), ref selectedMoveType);
                 moveInEditor.MoveType = (MoveType)selectedMoveType;
-                //file open for sprite... or should we enumerate them and display a list...?
-                ImGui.Text("Sprite ID");
-                ImGui.SameLine();
 
                 var spriteId = moveInEditor.SpriteName ?? string.Empty;
-                //var allSprites = _characterOperations.GetAllSprites(projectData.ProjectPathOnly);
-
                 var selectedSpriteIndex = spriteId != string.Empty ? allSprites.IndexOf(allSprites.First(x => x.Name == moveInEditor.SpriteName)) : -1;
-                ImGui.Combo("##SpriteSelection", ref selectedSpriteIndex, allSprites.Select(x => x.Name).ToArray(), allSprites.Count);
+                ImguiDrawingHelper.DrawComboInput("sprite", allSprites.Select(x => x.Name).ToArray(), ref selectedSpriteIndex);
+                
                 if (selectedSpriteIndex > -1)
                 {
                     if (moveInEditor.SpriteName != allSprites[selectedSpriteIndex].Name)
@@ -430,7 +432,11 @@ namespace CharacterDataEditor.Screens
                         ChangeAnimatedSprite(sprite);
                     }
                 }
-                
+
+                bool isThrow = moveInEditor.IsThrow;
+                ImguiDrawingHelper.DrawBoolInput("isMoveAThrow?", ref isThrow);
+                moveInEditor.IsThrow = isThrow;
+
                 if (ImGui.CollapsingHeader("Windows"))
                 {
                     int windowCount = moveInEditor.NumberOfFrames;
@@ -552,6 +558,8 @@ namespace CharacterDataEditor.Screens
                                 int particleDuration = attackDataItem.ParticleDuration;
                                 int holdOffsetX = attackDataItem.HoldXOffset;
                                 int holdOffsetY = attackDataItem.HoldYOffset;
+                                float comboScaling = attackDataItem.ComboScaling;
+                                float meterGain = attackDataItem.MeterGain;
 
                                 ImguiDrawingHelper.DrawIntInput("start", ref start);
                                 ImguiDrawingHelper.DrawIntInput("lifetime", ref lifetime);
@@ -561,6 +569,8 @@ namespace CharacterDataEditor.Screens
                                 var ySelect = ImguiDrawingHelper.DrawIntInput("heightOffset", ref heightOffset);
                                 ImguiDrawingHelper.DrawIntInput("group", ref group);
                                 ImguiDrawingHelper.DrawIntInput("damage", ref damage);
+                                ImguiDrawingHelper.DrawDecimalInput("meterGain", ref meterGain);
+                                ImguiDrawingHelper.DrawDecimalInput("comboScaling", ref comboScaling);
                                 ImguiDrawingHelper.DrawIntInput("attackHitStop", ref attackHitstop);
                                 ImguiDrawingHelper.DrawIntInput("attackHitStun", ref attackHitstun);
 
@@ -617,6 +627,8 @@ namespace CharacterDataEditor.Screens
                                 attackDataItem.ParticleDuration = particleDuration;
                                 attackDataItem.HoldXOffset = holdOffsetX;
                                 attackDataItem.HoldYOffset = holdOffsetY;
+                                attackDataItem.ComboScaling = comboScaling;
+                                attackDataItem.MeterGain = meterGain;
 
                                 ImGui.TreePop();
                             }
@@ -624,11 +636,131 @@ namespace CharacterDataEditor.Screens
                     }
                 }
 
-                bool isThrow = moveInEditor.IsThrow;
-                ImGui.Text("Is move a throw?");
-                ImGui.SameLine();
-                ImGui.Checkbox("##isThrow", ref isThrow);
-                moveInEditor.IsThrow = isThrow;
+                if (ImGui.CollapsingHeader("Rehit Data"))
+                {
+                    int rehitHitbox = moveInEditor.RehitData.HitBox;
+                    int numberOfRepeats = moveInEditor.RehitData.NumberOfHits;
+
+                    ImguiDrawingHelper.DrawIntInput("hitboxToRepeat", ref rehitHitbox, 0);
+                    ImguiDrawingHelper.DrawIntInput("numberOfRepeats", ref numberOfRepeats, 0);
+
+                    if (numberOfRepeats < 0)
+                    {
+                        numberOfRepeats = 0;
+                    }
+
+                    if (numberOfRepeats < moveInEditor.RehitData.NumberOfHits)
+                    {
+                        while (numberOfRepeats < moveInEditor.RehitData.NumberOfHits)
+                        {
+                            moveInEditor.RehitData.HitOnFrames.RemoveAt(moveInEditor.RehitData.NumberOfHits - 1);
+                        }
+                    }
+                    else
+                    {
+                        while (numberOfRepeats > moveInEditor.RehitData.NumberOfHits)
+                        {
+                            moveInEditor.RehitData.HitOnFrames.Add(0);
+                        }
+                    }
+
+                    if (numberOfRepeats == 0)
+                    {
+                        ImGui.Text("No Repeat Frame Information");
+                    }
+                    else
+                    {
+                        for (int i = 0; i < numberOfRepeats; i++)
+                        {
+                            var currentFrame = moveInEditor.RehitData.HitOnFrames[i];
+
+                            if (ImGui.TreeNode($"Hit On Frame [{i}]"))
+                            {
+                                ImguiDrawingHelper.DrawIntInput("repeatFrameIndex", ref currentFrame);
+
+                                ImGui.TreePop();
+                            }
+
+                            moveInEditor.RehitData.HitOnFrames[i] = currentFrame;
+                        }
+                    }
+
+                    moveInEditor.RehitData.HitBox = rehitHitbox;
+                }
+
+                if (isThrow && ImGui.CollapsingHeader("Opponent Position Data"))
+                {
+                    float distanceFromWall = moveInEditor.OpponentPositionData.DistanceFromWall;
+                    int numberOfFrames = moveInEditor.OpponentPositionData.NumberOfFrames;
+
+                    ImguiDrawingHelper.DrawDecimalInput("distanceFromWall", ref distanceFromWall);
+                    ImguiDrawingHelper.DrawIntInput("numberOfFrames", ref numberOfFrames);
+
+                    moveInEditor.OpponentPositionData.DistanceFromWall = distanceFromWall;
+
+                    if (numberOfFrames < 0)
+                    {
+                        numberOfFrames = 0;
+                    }
+
+                    if (numberOfFrames < moveInEditor.OpponentPositionData.NumberOfFrames)
+                    {
+                        while (numberOfFrames < moveInEditor.OpponentPositionData.NumberOfFrames)
+                        {
+                            moveInEditor.OpponentPositionData.Frames.RemoveAt(moveInEditor.OpponentPositionData.NumberOfFrames - 1);
+                        }
+                    }
+                    else
+                    {
+                        while (numberOfFrames > moveInEditor.OpponentPositionData.NumberOfFrames)
+                        {
+                            moveInEditor.OpponentPositionData.Frames.Add(new OpponentPositionFrameModel());
+                        }
+                    }
+
+                    if (numberOfFrames == 0)
+                    {
+                        ImGui.Text("No Frame Data");
+                    }
+                    else
+                    {
+                        for (int i = 0; i < numberOfFrames; i++)
+                        {
+                            var currentFrame = moveInEditor.OpponentPositionData.Frames[i];
+
+                            if (ImGui.TreeNode($"Frame [{i}]"))
+                            {
+                                int frameNumber = currentFrame.Frame;
+                                int relativeXPos = currentFrame.RelativeX;
+                                int relativeYPos = currentFrame.RelativeY;
+                                int frameIndex = currentFrame.Index;
+                                int rotation = currentFrame.Rotation;
+                                int xScale = currentFrame.XScale;
+                                int spriteToUse = (int)currentFrame.Sprite;
+
+                                ImguiDrawingHelper.DrawIntInput("frame", ref frameNumber);
+                                ImguiDrawingHelper.DrawIntInput("relativeXPosition", ref relativeXPos);
+                                ImguiDrawingHelper.DrawIntInput("relativeYPosition", ref relativeYPos);
+                                ImguiDrawingHelper.DrawComboInput("spriteToUse", spriteTypesList.ToArray(), ref spriteToUse);
+                                ImguiDrawingHelper.DrawIntInput("index", ref frameIndex);
+                                ImguiDrawingHelper.DrawIntInput("rotation", ref rotation);
+                                ImguiDrawingHelper.DrawIntInput("xScale", ref xScale);
+
+                                currentFrame.Frame = frameNumber;
+                                currentFrame.RelativeY = relativeYPos;
+                                currentFrame.RelativeX = relativeXPos;
+                                currentFrame.Index = frameIndex;
+                                currentFrame.Rotation = rotation;
+                                currentFrame.XScale = xScale;
+                                currentFrame.Sprite = (SpriteType)spriteToUse;
+
+                                ImGui.TreePop();
+                            }
+
+                            moveInEditor.OpponentPositionData.Frames[i] = currentFrame;
+                        }
+                    }
+                }
 
                 if (ImGui.CollapsingHeader("Hurtboxes"))
                 {
@@ -698,6 +830,60 @@ namespace CharacterDataEditor.Screens
 
                                 ImGui.TreePop();
                             }
+                        }
+                    }
+                }
+
+                if (ImGui.CollapsingHeader("Movement Data Frames"))
+                {
+                    int movementDataCount = moveInEditor.NumberOfMovementData;
+
+                    ImguiDrawingHelper.DrawIntInput("numberOfMovementDataFrames", ref movementDataCount);
+
+                    if (movementDataCount < 0)
+                    {
+                        movementDataCount = 0;
+                    }
+
+                    if (movementDataCount < moveInEditor.NumberOfMovementData)
+                    {
+                        while (movementDataCount < moveInEditor.NumberOfMovementData)
+                        {
+                            moveInEditor.MovementData.RemoveAt(moveInEditor.NumberOfMovementData - 1);
+                        }
+                    }
+                    else
+                    {
+                        while (movementDataCount > moveInEditor.NumberOfMovementData)
+                        {
+                            moveInEditor.MovementData.Add(new MovementDataModel());
+                        }
+                    }
+
+                    if (movementDataCount == 0)
+                    {
+                        ImGui.Text("No movement data frames");
+                    }
+                    else
+                    {
+                        for (int i = 0; i < moveInEditor.NumberOfMovementData; i++)
+                        {
+                            var currentMovementData = moveInEditor.MovementData[i];
+
+                            int startFrame = currentMovementData.StartingFrame;
+                            float horizontalSpeed = currentMovementData.HorizontalSpeed;
+                            float verticalSpeed = currentMovementData.VerticalSpeed;
+                            bool overwriteSpeed = currentMovementData.OverwriteSpeed;
+
+                            ImguiDrawingHelper.DrawIntInput("startingFrame", ref startFrame);
+                            ImguiDrawingHelper.DrawDecimalInput("horizontalSpeed", ref horizontalSpeed);
+                            ImguiDrawingHelper.DrawDecimalInput("verticalSpeed", ref verticalSpeed);
+                            ImguiDrawingHelper.DrawBoolInput("overwriteSpeed", ref overwriteSpeed);
+
+                            currentMovementData.StartingFrame = startFrame;
+                            currentMovementData.HorizontalSpeed = horizontalSpeed;
+                            currentMovementData.VerticalSpeed = verticalSpeed;
+                            currentMovementData.OverwriteSpeed = overwriteSpeed;
                         }
                     }
                 }
@@ -1053,7 +1239,7 @@ namespace CharacterDataEditor.Screens
                 if (ImGui.MenuItem("Save"))
                 {
                     //exports json to the characterdata folder in the game
-                    _characterOperations.SaveCharacter(character, projectData.ProjectPathOnly);
+                    SaveCharacter();
                 }
 
                 if (ImGui.MenuItem("Export..."))
@@ -1090,7 +1276,7 @@ namespace CharacterDataEditor.Screens
                         {
                             if (keycode == (int)KeyboardKey.KEY_S)
                             {
-                                _characterOperations.SaveCharacter(character, projectData.ProjectPathOnly);
+                                SaveCharacter();
                                 screenManager.NavigateTo(typeof(ProjectHomeScreen), new { height, width, projectData });
                             }
                             else if (keycode == (int)KeyboardKey.KEY_X)
@@ -1122,7 +1308,7 @@ namespace CharacterDataEditor.Screens
                         {
                             if (keycode == (int)KeyboardKey.KEY_S)
                             {
-                                _characterOperations.SaveCharacter(character, projectData.ProjectPathOnly);
+                                SaveCharacter();
                                 screenManager.NavigateTo(typeof(MainScreen), new { height, width });
                             }
                             else if (keycode == (int)KeyboardKey.KEY_X)
@@ -1155,7 +1341,7 @@ namespace CharacterDataEditor.Screens
                         {
                             if (keycode == (int)KeyboardKey.KEY_S)
                             {
-                                _characterOperations.SaveCharacter(character, projectData.ProjectPathOnly);
+                                SaveCharacter();
                                 screenManager.ExitWindow = true;
                             }
                             else if (keycode == (int)KeyboardKey.KEY_X)
@@ -1192,7 +1378,7 @@ namespace CharacterDataEditor.Screens
                 {
                     if (keycode == (int)KeyboardKey.KEY_S)
                     {
-                        _characterOperations.SaveCharacter(character, projectData.ProjectPathOnly);
+                        SaveCharacter();
                         screenManager.ExitWindow = true;
                     }
                     else if (keycode == (int)KeyboardKey.KEY_X)
@@ -1211,6 +1397,12 @@ namespace CharacterDataEditor.Screens
             {
                 screenManager.ExitWindow = true;
             }
+        }
+
+        private void SaveCharacter()
+        {
+            _characterOperations.SaveCharacter(character, projectData.ProjectPathOnly);
+            originalCharacter = character.Clone();
         }
     }
 }